refactor to add func_conversion
